+++ src/api/AbstractCRUDApi.ts
@@ -4,7 +4,7 @@ import * as querystring from "querystring";
 
 export default abstract class AbstractCRUDApi implements CRUDApi {
 
-    fetchAll(orderBy: string, orderWay: string, page: number, perPage: number): Promise<String> {
+    fetchItems(orderBy: string, orderWay: string, page: number, perPage: number): Promise<String> {
         let qs = querystring.stringify({
             sort: orderBy,
             direction: orderWay,
@@ -14,7 +14,7 @@ export default abstract class AbstractCRUDApi implements CRUDApi {
         return api.get(`${this.getBaseURL()}?${qs}`).then(res => res.data);
     }
 
-    fetchSingle(id: String): Promise<Object> {
+    fetchItem(id: String): Promise<Object> {
         return api.get(`${this.getBaseURL()}/${id}`).then(res => res.data);
     }