+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 + Add single fetch. Prepare for call detail
 + Add open close modal
 + Refactor actions
++ Refactor fetchItems and fetchItem
 
 ## 0.16.0
 + Add dashboard